Feat: 생성/수정 시간 자동 관리를 위한 BaseEntity 추가
- 각 entity 별로 createdAt, modifiedAt을 만들게 되면 비슷한 코드가 불필요하게 반복됨
- 따라서,  추상 클래스인 BaseEntity를 만들어 createdAt, modifiedAt을 관리하고 각 entity는 BaseEntity를 extend해서 사용하게끔 수정
- main에 JPA Auditing을 활성화 하기 위한 어노테이션 추가
