+++ src/main/java/grabit/grabit_backend/Domain/Challenge.java
@@ -1,15 +1,12 @@
 package grabit.grabit_backend.Domain;
 
-import org.apache.tomcat.jni.Local;
-
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-import java.time.LocalDateTime;
 
 @Entity(name = "challenge")
-public class Challenge {
+public class Challenge extends BaseEntity{
 
 	@Id @GeneratedValue
 	@Column(name = "CHALLENGE_ID")
@@ -18,16 +15,12 @@ public class Challenge {
 	@Column(name = "LEADER_ID")
 	private String leaderId;
 	private String name;
-	private LocalDateTime createdAt;
-	private LocalDateTime modifiedAt;
 
 	public Challenge(){}
 
 	public Challenge(String leaderId, String name) {
 		this.leaderId = leaderId;
 		this.name = name;
-		this.createdAt = LocalDateTime.now();
-		this.modifiedAt = LocalDateTime.now();
 	}
 
 	public Long getId() {
@@ -54,19 +47,4 @@ public class Challenge {
 		this.name = name;
 	}
 
-	public LocalDateTime getCreatedAt() {
-		return createdAt;
-	}
-
-	public void setCreatedAt(LocalDateTime createdAt) {
-		this.createdAt = createdAt;
-	}
-
-	public LocalDateTime getModifiedAt() {
-		return modifiedAt;
-	}
-
-	public void setModifiedAt(LocalDateTime modifiedAt) {
-		this.modifiedAt = modifiedAt;
-	}
 }
